requirements.txt
# required packages

requests==2.26.0
beautifulsoup4==4.10.0
python-dotenv==0.19.2
transformers==4.11.3
gensim==4.1.2
streamlit==0.89.0